adding Thermometer and starting to use it
[git-p4: depot-paths = "//depot/ACG/Smith/Firmware/": change = 97908]


Former-commit-id: 9094b96b7e919153526f0139e31d8150bb4d0b47

+++ C++/include/PrinterStatus.h
@@ -94,7 +94,7 @@ public:
     int _numLayers;
     int _currentLayer;
     int _estimatedSecondsRemaining;
-    float _temperature;
+    double _temperature;
 };
 
 #endif	/* PRINTERSTATUS_H */
